fix build

+++ src/Graphics/Meshes/MeshRenderer.cpp
@@ -1,4 +1,7 @@
 #include "MeshRenderer.h"
+
+#include <gsl/gsl>
+
 namespace trg
 {
 	namespace MeshRendererPrivate
@@ -45,7 +48,7 @@ namespace trg
 		m_vertexBuffer.bind(bindInfo);
 		m_indexBuffer.bind(bindInfo);
 
-		bindInfo.commandBuffer->drawIndexed(m_meshFilter.m_mesh.getIndexData().size(),
+		bindInfo.commandBuffer->drawIndexed(gsl::narrow<uint32_t>(m_meshFilter.m_mesh.getIndexData().size()),
 											1 /*instanceCount*/,
 											0 /*firstIndex*/,
 											0 /*vertexOffset*/,
